Move and refactor metadatadate logic from IngridIndexExporter to IngridIDFExporter
