+++ server/src/main/java/de/ingrid/igeserver/profiles/ingrid/exporter/IngridModelTransformer.kt
@@ -777,7 +777,7 @@ open class IngridModelTransformer(
 
     val parentIdentifier: String? = data.parentIdentifier
     val hierarchyParent: String? = data._parent
-    val modifiedMetadataDate: String = formatDate(formatterOnlyDate, data.modifiedMetadata ?: model._contentModified)
+
     var pointOfContact: List<AddressModelTransformer> = emptyList()
     var orderInfoContact: List<AddressModelTransformer>
     fun getAddressesToUuids() = pointOfContact.flatMap { model ->
