+++ server/src/test/kotlin/de/ingrid/igeserver/exports/ingrid/Geodataservice.kt
@@ -67,9 +67,9 @@ class Geodataservice : ShouldSpec() {
 
     override suspend fun beforeSpec(spec: Spec) {
         clearAllMocks()
-        this.exporter = IngridIDFExporter(codelistHandler, uploadConfig, catalogService, documentService)
+        this.exporter = IngridIDFExporter(codelistHandler, uploadConfig, catalogService, documentService, documentWrapperRepository)
         this.luceneExporter = IngridLuceneExporter(codelistHandler, uploadConfig, catalogService, documentService)
-        this.indexExporter = IngridIndexExporter(this.exporter, this.luceneExporter, documentWrapperRepository)
+        this.indexExporter = IngridIndexExporter(this.exporter, this.luceneExporter)
 
         mockkObject(SpringContext.Companion)
         every { SpringContext.getBean(DocumentService::class.java) } answers {
