+++ server/src/test/kotlin/de/ingrid/igeserver/exports/ingrid_bast/BastFields.kt
@@ -45,6 +45,7 @@ class BastFields : GeodatasetBase() {
                 this.uploadConfig,
                 this.catalogService,
                 this.documentService,
+                this.documentWrapperRepository,
             )
         every { catalogService.getProfileFromCatalog(any()) } returns
             DummyCatalog("ingrid-bast")
