+++ server/src/test/kotlin/de/ingrid/igeserver/exports/ingrid_bast/BastFieldsExternal.kt
@@ -46,6 +46,7 @@ class BastFieldsExternal : GeodatasetBase() {
                 this.uploadConfig,
                 this.catalogService,
                 this.documentService,
+                this.documentWrapperRepository,
             )
         every { catalogService.getProfileFromCatalog(any()) } returns
             DummyCatalog("ingrid-bast")
