+++ server/src/test/kotlin/de/ingrid/igeserver/exports/ingrid_hmdk/HmdkFields.kt
@@ -48,6 +48,7 @@ class HmdkFields : GeodatasetBase() {
                 this.uploadConfig,
                 this.catalogService,
                 this.documentService,
+                this.documentWrapperRepository,
             )
         every { catalogService.getProfileFromCatalog(any()) } returns
             DummyCatalog("ingrid-hmdk")
