+++ server/src/test/kotlin/de/ingrid/igeserver/exports/ingrid_krzn/KrznFields.kt
@@ -41,6 +41,7 @@ class KrznFields : GeodatasetBase() {
                 this.uploadConfig,
                 this.catalogService,
                 this.documentService,
+                this.documentWrapperRepository,
             )
         every { catalogService.getProfileFromCatalog(any()) } returns
             DummyCatalog("ingrid-krzn")
