+++ server/src/test/kotlin/de/ingrid/igeserver/exports/ingrid_lubw/LubwFields.kt
@@ -43,6 +43,7 @@ class LubwFields : GeodatasetBase() {
                 this.uploadConfig,
                 this.catalogService,
                 this.documentService,
+                this.documentWrapperRepository,
             )
         every { catalogService.getProfileFromCatalog(any()) } returns
             DummyCatalog("ingrid-lubw")
