+++ server/src/test/kotlin/de/ingrid/igeserver/exports/ingrid_up_sh/GeometryContext.kt
@@ -39,6 +39,7 @@ class GeometryContext : GeodatasetBase() {
                 this.uploadConfig,
                 this.catalogService,
                 this.documentService,
+                this.documentWrapperRepository,
             )
         every { catalogService.getProfileFromCatalog(any()) } returns
             DummyCatalog("ingrid-up-sh")
